+++ src/Campaign.php
@@ -143,7 +143,7 @@ class Campaign extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '2.0.0-beta.3';
+    public string $schemaVersion = '2.0.0-beta.4';
 
     /**
      * @inheritdoc